requirements.txt
babel==2.16.0
cffi==1.17.0
contourpy==1.3.0
cycler==0.12.1
docxcompose==1.4.0
docxtpl==0.18.0
et-xmlfile==1.1.0
fonttools==4.53.1
Jinja2==3.1.4
kiwisolver==1.4.5
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.9.2
numpy==2.1.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pycparser==2.22
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-pptx==1.0.2
pytz==2024.1
pyvips==2.2.3
setuptools==74.0.0
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.1
XlsxWriter==3.2.0
